Re-structure the order of the project
Also, make it PreservOnRefresh Compatible
