+++ README.md
@@ -1,37 +1,110 @@
-[![Published on Vaadin  Directory](https://img.shields.io/badge/Vaadin%20Directory-published-00b4f0.svg)](https://vaadin.com/directory/component/maskedtextfield-for-vaadin-8)
-[![Stars on Vaadin Directory](https://img.shields.io/vaadin-directory/star/maskedtextfield-for-vaadin-8.svg)](https://vaadin.com/directory/component/maskedtextfield-for-vaadin-8)
+# MyComponent Add-on for Vaadin 8
 
-# MaskedTextField for Vaadin 8
+${ComponentClassName} is a UI component add-on for Vaadin 8.
 
-Vaadin 8 version of [MaskedTextField](http://vaadin.com/addon/maskedtextfield)
+## Online demo
 
+Try the add-on demo at <url of the online demo>
 
-# Usage
-Overview of currently available masks:
-```
-    # - any digit
-    U - upper-case letter
-    L - lower-case letter
-    ? - any letter
-    A - any number or character
-    * - anything
-    H - hex sign (0-9, a-f or A-F)
-    ' - Escape character, used to escape any of the special formatting characters.
-    ~ - +/- sign
-```
-Any character not matching one of the above mask character or if it escaped with the single quote character (') is considered to be a literal.
+## Download release
 
-Some mask examples:
-```
-    Phone Number: (###) ###-####
-    USPS Express Mail: EU#########'US
-    Date / time: ##/##/#### ##:##
-    State: UU
-    HTML Color: '#HHHHHH
-    An capitalized 6 letter word: ULLLLL
-```
+Official releases of this add-on are available at Vaadin Directory. For Maven instructions, download and reviews, go to http://vaadin.com/addon/maskedtextfield
 
-# Known issues:
+## Building and running demo
 
-- [Issues](https://github.com/andersonfreitas/vaadin-masked-textfield/issues)
+git clone <url of the MyComponent repository>
+mvn clean install
+cd demo
+mvn jetty:run
 
+To see the demo, navigate to http://localhost:8080/
+
+## Development with Eclipse IDE
+
+For further development of this add-on, the following tool-chain is recommended:
+- Eclipse IDE
+- m2e wtp plug-in (install it from Eclipse Marketplace)
+- Vaadin Eclipse plug-in (install it from Eclipse Marketplace)
+- JRebel Eclipse plug-in (install it from Eclipse Marketplace)
+- Chrome browser
+
+### Importing project
+
+Choose File > Import... > Existing Maven Projects
+
+Note that Eclipse may give "Plugin execution not covered by lifecycle configuration" errors for pom.xml. Use "Permanently mark goal resources in pom.xml as ignored in Eclipse build" quick-fix to mark these errors as permanently ignored in your project. Do not worry, the project still works fine. 
+
+### Debugging maskedtextfield.server-side
+
+If you have not already compiled the widgetset, do it now by running vaadin:install Maven target for maskedtextfield-root project.
+
+If you have a JRebel license, it makes on the fly code changes faster. Just add JRebel nature to your maskedtextfield-demo project by clicking project with right mouse button and choosing JRebel > Add JRebel Nature
+
+To debug project and make code modifications on the fly in the maskedtextfield.server-side, right-click the maskedtextfield-demo project and choose Debug As > Debug on Server. Navigate to http://localhost:8080/maskedtextfield-demo/ to see the application.
+
+### Debugging maskedtextfield.client-side
+
+Debugging maskedtextfield.client side code in the maskedtextfield-demo project:
+  - run "mvn vaadin:run-codeserver" on a separate console while the application is running
+  - activate Super Dev Mode in the debug window of the application or by adding ?superdevmode to the URL
+  - You can access Java-sources and set breakpoints inside Chrome if you enable source maps from inspector settings.
+ 
+## Release notes
+
+### Version 1.0-SNAPSHOT
+- ...
+- ...
+
+## Roadmap
+
+This component is developed as a hobby with no public roadmap or any guarantees of upcoming releases. That said, the following features are planned for upcoming releases:
+- ...
+- ...
+
+## Issue tracking
+
+The issues for this add-on are tracked on its github.com page. All bug reports and feature requests are appreciated. 
+
+## Contributions
+
+Contributions are welcome, but there are no guarantees that they are accepted as such. Process for contributing is the following:
+- Fork this project
+- Create an issue to this project about the contribution (bug or feature) if there is no such issue about it already. Try to keep the scope minimal.
+- Develop and test the fix or functionality carefully. Only include minimum amount of code needed to fix the issue.
+- Refer to the fixed issue in commit
+- Send a pull request for the original project
+- Comment on the original issue that you have implemented a fix for it
+
+## License & Author
+
+Add-on is distributed under Apache License 2.0. For license terms, see LICENSE.txt.
+
+MyComponent is written by <...>
+
+# Developer Guide
+
+## Getting started
+
+Here is a simple example on how to try out the add-on component:
+
+<...>
+
+For a more comprehensive example, see src/test/java/org/vaadin/template/demo/DemoUI.java
+
+## Features
+
+### Feature A
+
+<...>
+
+### Feature B
+
+<...>
+
+### Feature C
+
+<...>
+
+## API
+
+MyComponent JavaDoc is available online at <...>
