+++ maskedtextfield-addon/src/main/java/org/vaadin/addons/maskedtextfield/client/DecimalFieldState.java
