+++ maskedtextfield-addon/src/main/java/org/vaadin/addons/maskedtextfield/client/PrefixedMaskedTextFieldState.java
