+++ maskedtextfield-addon/src/main/java/org/vaadin/addons/maskedtextfield/client/masks/Mask.java
