+++ maskedtextfield-addon/src/main/resources/org/vaadin/addons/public/maskedtextfield/styles.css
@@ -1,0 +1,7 @@
+div.maskedtextfield {
+	cursor: pointer;
+	user-select: none;
+	-webkit-user-select: none;
+	-moz-user-select: none;
+	-ms-user-select: none;
+}
