--- src/main/java/org/vaadin/addons/maskedtextfield/MaskedTextFieldWidgetset.gwt.xml
@@ -1,12 +1,0 @@
-<module>
-
-	<!-- Inherit DefaultWidgetSet -->
-	<inherits name="com.vaadin.DefaultWidgetSet" />
-
-	<set-configuration-property name="devModeRedirectEnabled" value="true" />
-	<add-linker name="xsiframe" />
-
-	<source path="client" />
-	<source path="shared" />
-
-</module>
