--- template.mf
@@ -1,11 +1,0 @@
-Bundle-ManifestVersion: 2
-Bundle-SymbolicName: org.vaadin.addons.maskedtextfield
-Bundle-Name: Masked Text Field for Vaadin 8
-Bundle-Vendor: Vaadin Community
-Excluded-Imports: 
- com.google.gwt*,
- com.vaadin.client.*
-Excluded-Exports: 
- org.vaadin.addons.maskedtextfield.client.*
-Import-Template:
- com.vaadin.*;version="[8.4.4, 9.0.0)"
